+++ src/components/code/Code.jsx
@@ -1,6 +1,6 @@
 import React, {useEffect, useState} from "react";
 import RegistryView from "./RegistryView.jsx";
-import {loadWasm} from "../rust_functions.js";
+import {loadWasm} from "../../rust_functions.js";
 
 import MemoryView from "./MemoryView.jsx";
 import Console from "./Console.jsx";
