+++ src/components/code/MemoryView.jsx
@@ -1,5 +1,5 @@
 import React, {useCallback, useEffect, useRef, useState} from "react";
-import {RUST} from "../rust_functions.js";
+import {RUST} from "../../rust_functions.js";
 import {CALLBACK_TYPES, CALLBACKS_TRIGGERS} from "./simulator.js";
 
 const WIDTH = 4;
