+++ src/components/code/RegistryView.jsx
@@ -1,5 +1,5 @@
 import React, {useEffect, useState} from "react";
-import {RUST} from "../rust_functions.js";
+import {RUST} from "../../rust_functions.js";
 import {CALLBACK_TYPES, CALLBACKS_TRIGGERS} from "./simulator.js";
 
 function RegistryView({loaded, registerCallback}) {
